Snowhead/Great Bay logic
This should fix up all of the logic in Snowhead Temple and in the Great Bay region in general, including in Great Bay Temple. For now I have slapped a hard Hookshot requirement on all GBT spots, but in theory Hookshot is only hard required for the boss (and either it or ice arrows are needed for the Map). The reasoning here is that GBT is impossible to leave without Hookshot (or defeating the boss... which requires Hookshot) so if time sensitive items land in here the player would be screwed without Hookshot. I hope to have a better solution for this in the future.