Logic fixes, start on Ikana
This fixes a lot of logic flaws I overlooked in previous work and does most of the Ikana region. In theory all the logic should be correct now outside of everything involving the Anju/Kafei sidequest and the actual interior of Stone Tower Temple (both normal and inverted), including the Great Fairy Sword slot.