Logic can define forbidden placements for items
Adds a section in the bottom of logic that can list places where an
item should never be placed. If there is no section is defined it falls
back to the hard-coded placements we have now.

The section should always start with the header:

--Forbidden Item Placements--

The forbidden locations can be formatted as a list:

-Mama_Letter:Bomb_Bag,Quiver_1,Bomb_Bag_1

Or can be defined as ranges:

-Blast_Mask:Moon_Tear-Pendant

Or any combination of the two:

-Keaton_Mask:M_Shield,Wallet_2-Mama_Letter

They can also be spread out over multiple lines by just repeating the
key again

-Bottle_G:Hero_Bow-Light_Arrow,Lens,Hookshot,Bottle_D,M_Shield
-Bottle_G:Captain_Hat,Giant_Mask,WF_Map-ST_Key4

There is currently no UI for adding this information and therefore any
logic edits done through the editor will currently delete this section.